Add compiler built-in sanitizer
See https://releases.llvm.org/13.0.0/tools/clang/docs/LeakSanitizer.html
for details.

Valgrind doesn't get along with binaries compiled with sanitizers, and
also it's not a good idea to use such a binary outside of developer's
machine testing, so there are make targets to compile with and without
this option.