Add resolver
This doesn't cover bad cases, only the good path.

Some expressions are now marked as `frozen`. That was done
so as to make these types hashable and hence usable as keys
in dictionaries. See dataclass documentation for more info.

Token became not just `frozen` (for the same reason),
but it now doesn't implement `eq`. Without it if a variable
on a single line has different scopes (either all the source
is a single line or in a `for` loop), then Python doesn't
see any difference between all those occurrences and
newer ones override older ones in the dict.
In original Java interpreter this class doesn't have a custom
`equals` method, so equality is based on hashes, which are
different if links to the objects are different. So in Java
there's no need to explicitly tell "no equality". In Python
it also wouldn't be necessary by default, but `dataclass`
has `eq` enabled, so it has to be turned off explicitly.